Add testcase to demonstrate headers are not per-instance
CastorClient.headers is defined as a class attribute rather than an
instance attribute. This is a problem because authorization updates
self.headers to set the bearer token, which means that if a user were to
create a second instance of CastorClient, it would overwrite/redefine
the token used by the first client.

This test demonstrates the problem, to provide assurance that the next
commit in this tree fixes the issue.