Prevent instances of CastorClient overwriting each other's tokens
As described in 8fe56fe0dbc7b6dd1d5fa9216b76a17c99debed4,
CastorClient.headers is defined as a class attribute rather than an
instance attribute. This is a problem because authorization updates
self.headers to set the bearer token, which means that if a user were to
create a second instance of CastorClient, it would overwrite/redefine
the token used by the first client.

By changing headers from a class attribute to an instance attribute, the
problem is avoided.